Split config and write config file before launching the dev server
