+++ docs/.vuepress/config.js
@@ -538,7 +538,7 @@ const sidebar = {
   ],
 };
 
-  
+
 const checklinksIgnoredFiles = [
   '**/node_modules', // please never remove this one
   /**
@@ -557,6 +557,7 @@ const checklinksIgnoredFiles = [
   './developer-docs/latest/setup-deployment-guides/configurations.md', // the script thinks filename[]() at line 977 is a real link
   './developer-docs/latest/development/backend-customization/webhooks.md', // 'missing" links are in commented part of file
 ];
+
 
 const plugins = [
   ['vuepress-plugin-element-tabs', {}],
@@ -663,12 +664,12 @@ const plugins = [
   ],
 ];
 
-module.exports = {
+
+const metas = {
   title: '',
   port: 8080,
   description: 'The headless CMS developers love.',
   base: '/',
-  plugins: plugins,
   head: [
     [
       'link',
@@ -765,175 +766,178 @@ module.exports = {
     [
       'script',
       {},
-      `(function(w, d, s, l, i) {
-  w[l] = w[l] || [];
-  w[l].push({ 'gtm.start': new Date().getTime(), event: 'gtm.js' });
-  var f = d.getElementsByTagName(s)[0],
-    j = d.createElement(s),
-    dl = l != 'dataLayer' ? '&l=' + l : '';
-  j.async = true;
-  j.src = 'https://www.googletagmanager.com/gtm.js?id=' + i + dl;
-  f.parentNode.insertBefore(j, f);
-})(window, document, 'script', 'dataLayer', 'GTM-KN9JRWG');
-`,
+      `(function(w,d,s,l,i){w[l]=w[l]||[];w[l].push({'gtm.start': new Date().getTime(),event:'gtm.js'});var f=d.getElementsByTagName(s)[0], j=d.createElement(s),dl=l!='dataLayer'?'&l='+l:'';j.async=true;j.src= 'https://www.googletagmanager.com/gtm.js?id='+i+dl;f.parentNode.insertBefore(j,f); })(window,document,'script','dataLayer','GTM-KN9JRWG');`,
     ],
   ],
-  themeConfig: {
-    logo: '/assets/logo.png',
-    nav: [
-      {
-        text: 'Resource Center',
-        link: 'https://strapi.io/resource-center',
-      },
-      {
-        text: 'v4 Documentation',
-        items: [
-          {
-            text: 'Developer Docs',
-            items: [
-              {
-                text: 'Getting Started',
-                link: '/developer-docs/latest/getting-started/introduction.html',
-              },
-              {
-                text: 'Setup & Deployment',
-                link: '/developer-docs/latest/setup-deployment-guides/installation.html',
-              },
-              {
-                text: 'Plugins',
-                link: '/developer-docs/latest/plugins/plugins-intro.html',
-              },
-              {
-                text: 'Development',
-                link: '/developer-docs/latest/development/backend-customization.html',
-              },
-              {
-                text: 'Update & Migration',
-                link: '/developer-docs/latest/update-migration-guides/update-version.html',
-              },
-              {
-                text: 'Developer Resources',
-                link: '/developer-docs/latest/developer-resources/content-api/content-api.html',
-              },
-            ],
-          },
-          {
-            text: 'User Guide',
-            items: [
-              {
-                text: 'Getting Started',
-                link: '/user-docs/latest/getting-started/introduction.html',
-              },
-              {
-                text: 'Content Manager',
-                link: '/user-docs/latest/content-manager/introduction-to-content-manager.html',
-              },
-              {
-                text: 'Content-Type Builder',
-                link:
-                  '/user-docs/latest/content-types-builder/introduction-to-content-types-builder.html',
-              },
-              {
-                text: 'Users, Roles, and Permissions',
-                link:
-                  '/user-docs/latest/users-roles-permissions/introduction-to-users-roles-permissions.html',
-              },
-              {
-                text: 'Plugins',
-                link: '/user-docs/latest/plugins/introduction-to-plugins.html',
-              },
-              {
-                text: 'General Settings',
-                link: '/user-docs/latest/settings/managing-global-settings.html',
-              },
-            ],
-          },
-        ],
-      },
-      {
-        text: 'v3 documentation',
-        link: 'https://docs-v3.strapi.io'
-      },
-      {
-        text: 'Ecosystem',
-        items: [
-          {
-            text: 'Strapi',
-            items: [
-              {
-                text: 'Website',
-                link: 'https://strapi.io',
-              },
-              {
-                text: 'Blog',
-                link: 'https://strapi.io/blog',
-              },
-              {
-                text: 'StrapiConf 2021',
-                link: 'https://www.strapi.io/strapi-conf-2021',
-              },
-            ],
-          },
-          {
-            text: 'Community',
-            items: [
-              {
-                text: 'Forum',
-                link: 'https://forum.strapi.io',
-              },
-              {
-                text: 'Discord',
-                link: 'https://discord.strapi.io',
-              },
-              {
-                text: 'Awesome-Strapi',
-                link: 'https://github.com/strapi/awesome-strapi',
-              },
-            ],
-          },
-          {
-            text: 'Resources',
-            items: [
-              {
-                text: 'Tutorials',
-                link: 'https://strapi.io/tutorials',
-              },
-              {
-                text: 'Academy',
-                link: 'https://academy.strapi.io/',
-              },
-            ],
-          },
-        ],
-      },
-      {
-        text: "We're hiring!",
-        link: 'https://strapi.io/careers#open-positions',
-      },
-    ],
-    repo: 'strapi/documentation',
-    docsDir: 'docs',
-    docsBranch: 'main',
-    algolia: {
-      appId: '9FTY6J9E4X',
-      apiKey: 'cf49c82a1865df2618a3d89e18657051',
-      indexName: 'documentation',
+};
+;
+
+const themeConfig = {
+  logo: '/assets/logo.png',
+  nav: [
+    {
+      text: 'Resource Center',
+      link: 'https://strapi.io/resource-center',
     },
-    editLinks: true,
-    editLinkText: 'Improve this page',
-    serviceWorker: true,
-    sidebarDepth: 1,
-    smoothScroll: false,
-    sidebar: {
-      '/developer-docs/latest/': sidebar.developer,
-      '/user-docs/latest/': sidebar.user,
+    {
+      text: 'v4 Documentation',
+      items: [
+        {
+          text: 'Developer Docs',
+          items: [
+            {
+              text: 'Getting Started',
+              link: '/developer-docs/latest/getting-started/introduction.html',
+            },
+            {
+              text: 'Setup & Deployment',
+              link: '/developer-docs/latest/setup-deployment-guides/installation.html',
+            },
+            {
+              text: 'Plugins',
+              link: '/developer-docs/latest/plugins/plugins-intro.html',
+            },
+            {
+              text: 'Development',
+              link: '/developer-docs/latest/development/backend-customization.html',
+            },
+            {
+              text: 'Update & Migration',
+              link: '/developer-docs/latest/update-migration-guides/update-version.html',
+            },
+            {
+              text: 'Developer Resources',
+              link: '/developer-docs/latest/developer-resources/content-api/content-api.html',
+            },
+          ],
+        },
+        {
+          text: 'User Guide',
+          items: [
+            {
+              text: 'Getting Started',
+              link: '/user-docs/latest/getting-started/introduction.html',
+            },
+            {
+              text: 'Content Manager',
+              link: '/user-docs/latest/content-manager/introduction-to-content-manager.html',
+            },
+            {
+              text: 'Content-Type Builder',
+              link:
+                '/user-docs/latest/content-types-builder/introduction-to-content-types-builder.html',
+            },
+            {
+              text: 'Users, Roles, and Permissions',
+              link:
+                '/user-docs/latest/users-roles-permissions/introduction-to-users-roles-permissions.html',
+            },
+            {
+              text: 'Plugins',
+              link: '/user-docs/latest/plugins/introduction-to-plugins.html',
+            },
+            {
+              text: 'General Settings',
+              link: '/user-docs/latest/settings/managing-global-settings.html',
+            },
+          ],
+        },
+      ],
     },
+    {
+      text: 'v3 documentation',
+      link: 'https://docs-v3.strapi.io',
+    },
+    {
+      text: 'Ecosystem',
+      items: [
+        {
+          text: 'Strapi',
+          items: [
+            {
+              text: 'Website',
+              link: 'https://strapi.io',
+            },
+            {
+              text: 'Blog',
+              link: 'https://strapi.io/blog',
+            },
+            {
+              text: 'StrapiConf 2021',
+              link: 'https://www.strapi.io/strapi-conf-2021',
+            },
+          ],
+        },
+        {
+          text: 'Community',
+          items: [
+            {
+              text: 'Forum',
+              link: 'https://forum.strapi.io',
+            },
+            {
+              text: 'Discord',
+              link: 'https://discord.strapi.io',
+            },
+            {
+              text: 'Awesome-Strapi',
+              link: 'https://github.com/strapi/awesome-strapi',
+            },
+          ],
+        },
+        {
+          text: 'Resources',
+          items: [
+            {
+              text: 'Tutorials',
+              link: 'https://strapi.io/tutorials',
+            },
+            {
+              text: 'Academy',
+              link: 'https://academy.strapi.io/',
+            },
+          ],
+        },
+      ],
+    },
+    {
+      text: "We're hiring!",
+      link: 'https://strapi.io/careers#open-positions',
+    },
+  ],
+  repo: 'strapi/documentation',
+  docsDir: 'docs',
+  docsBranch: 'main',
+  algolia: {
+    appId: '9FTY6J9E4X',
+    apiKey: 'cf49c82a1865df2618a3d89e18657051',
+    indexName: 'documentation',
   },
-  markdown: {
-    extendMarkdown: md => {
-      // use more markdown-it plugins!
-      md.use(require('markdown-it-include'));
-    },
+  editLinks: true,
+  editLinkText: 'Improve this page',
+  serviceWorker: true,
+  sidebarDepth: 1,
+  smoothScroll: false,
+  sidebar: {
+    '/developer-docs/latest/': sidebar.developer,
+    '/user-docs/latest/': sidebar.user,
   },
+};
+
+
+const markdown = {
+  extendMarkdown: md => {
+    // use more markdown-it plugins!
+    md.use(require('markdown-it-include'));
+  },
+};
+
+
+
+module.exports = {
+  ...metas,
+  themeConfig,
+  markdown,
+  plugins,
 };
