+++ docs/.vuepress/markdown-config.js
@@ -1,0 +1,6 @@
+const markdown = {
+  extendMarkdown: md => {
+    // use more markdown-it plugins!
+    md.use(require('markdown-it-include'));
+  },
+};
