+++ docs/.vuepress/metas.js
@@ -1,0 +1,105 @@
+const metas = {
+  title: '',
+  port: 8080,
+  description: 'The headless CMS developers love.',
+  base: '/',
+  head: [
+    [
+      'link',
+      {
+        rel: 'icon',
+        href: 'https://strapi.io/assets/favicon-32x32.png',
+      },
+    ],
+    [
+      'meta',
+      {
+        property: 'og:title',
+        content: 'Strapi Documentation',
+      },
+    ],
+    [
+      'meta',
+      {
+        property: 'og:type',
+        content: 'article',
+      },
+    ],
+    [
+      'meta',
+      {
+        property: 'og:url',
+        content: 'https://strapi.io/documentation/',
+      },
+    ],
+    [
+      'meta',
+      {
+        property: 'og:description',
+        content: 'The headless CMS developers love.',
+      },
+    ],
+    [
+      'meta',
+      {
+        property: 'og:image',
+        content: 'https://strapi.io/documentation/assets/meta.png',
+      },
+    ],
+    [
+      'meta',
+      {
+        property: 'og:article:author',
+        content: 'strapi',
+      },
+    ],
+
+    [
+      'meta',
+      {
+        property: 'twitter:card',
+        content: 'summary_large_image',
+      },
+    ],
+    [
+      'meta',
+      {
+        property: 'twitter:url',
+        content: 'https://strapi.io/documentation/',
+      },
+    ],
+    [
+      'meta',
+      {
+        property: 'twitter:site',
+        content: '@strapijs',
+      },
+    ],
+    [
+      'meta',
+      {
+        property: 'twitter:title',
+        content: 'Strapi Documentation',
+      },
+    ],
+    [
+      'meta',
+      {
+        property: 'twitter:description',
+        content: 'The headless CMS developers love.',
+      },
+    ],
+    [
+      'meta',
+      {
+        property: 'twitter:image',
+        content: 'http://strapi.io/assets/images/strapi-website-preview.png',
+      },
+    ],
+    [
+      'script',
+      {},
+      `(function(w,d,s,l,i){w[l]=w[l]||[];w[l].push({'gtm.start': new Date().getTime(),event:'gtm.js'});var f=d.getElementsByTagName(s)[0], j=d.createElement(s),dl=l!='dataLayer'?'&l='+l:'';j.async=true;j.src= 'https://www.googletagmanager.com/gtm.js?id='+i+dl;f.parentNode.insertBefore(j,f); })(window,document,'script','dataLayer','GTM-KN9JRWG');`,
+    ],
+  ],
+};
